pyproject.toml
[tool.poetry]
name = "meetingroommanagement"
version = "0.1.0"
description = ""
authors = ["Pedro da Matta <pmrm@cin.ufpe.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
black = "^21.5b1"
flake8 = "^3.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
